Switch to using bigDecimal for more accurate calculations. Change type of discount factor to real to store 1/3 more accurately.

+++ src/main/java/com/example/checkout/CheckoutController.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -19,6 +20,6 @@ public class CheckoutController {
 
     @PostMapping("checkout")
     public float checkout(@RequestBody List<String> watchIds) {
-        return this.checkoutService.calculateTotalPrice(watchIds);
+        return this.checkoutService.calculateTotalPrice(watchIds).floatValue();
     }
 }
